Petzval lens working again
