+++ www/js/modules/planoconvex_lens.js
@@ -1,19 +1,24 @@
-export const surfaces = [
-    {"ObjectPlane": {"diam": 50}},
-    {"RefractingCircularConic": {"diam": 25.0, "roc": 25.8, "k": 0.0}},
-    {"RefractingCircularFlat": {"diam": 25.0}},
-    {"ImagePlane": {"diam": 50}},
+const AIR = {"real": {"Constant": 1.0}, "imag": null}
+const NBK7 = {"real": {"Constant": 1.515}, "imag": null}
+
+export const surface_specs = [
+    "Object",
+    {"Conic": {"semi_diameter": 12.5, "radius_of_curvature": 25.8, "conic_constant": 0.0, "surf_type": "Refracting"}},
+    {"Conic": {"semi_diameter": 12.5, "radius_of_curvature": Infinity, "conic_constant": 0.0, "surf_type": "Refracting"}},
+    "Image",
 ]
 
-export const gaps = [
-    {"n": 1.0, "thickness": Infinity},
-    {"n": 1.515, "thickness": 5.3},
-    {"n": 1.0, "thickness": 46.6},
+export const gap_specs = [
+    {"thickness": Infinity, "refractive_index": AIR},
+    {"thickness": 5.3, "refractive_index": NBK7},
+    {"thickness": 46.6, "refractive_index": AIR},
 ]
 
-export const aperture = {"EntrancePupilDiameter": { diam: 10 }}
+export const aperture_spec = {"EntrancePupil": { "semi_diameter": 5.0 }}
 
-export const fields = [
-    {"Angle": {"angle": 0, "wavelength": 0.5876, "sampling": {"SqGrid": {"spacing": 0.1}}}},
-    {"Angle": {"angle": 5, "wavelength": 0.5876, "sampling": {"SqGrid": {"spacing": 0.1}}}}
+export const field_specs = [
+    {"Angle": {"angle": 0, "pupil_sampling": {"SquareGrid": {"spacing": 0.1}}}},
+    {"Angle": {"angle": 5, "pupil_sampling": {"SquareGrid": {"spacing": 0.1}}}}
 ]
+
+export const wavelengths = [0.5876]
